Extract 'TraitDeclParser'
